requirements.txt
torch
numpy
yacs
torchvision
pillow
pandas
scipy
tqdm
scikit-learn